update headless dll path in extension to point to git submodule
This will only ever work for local/debug builds and release builds have not yet been accounted for. Why? Well I don't know the best way to publish the C# code with the extension. I also dont know yet whether it is worth going to the trouble of doing something fancy - or if I should just slap an xcopy at the end of the yarn build